refactor: Initial project updates
