+++ examples/code-only/Example01_Basic2DScene/Example01_Basic2DScene.csproj
@@ -12,6 +12,7 @@
     </Target>
 
     <ItemGroup>
+      <ProjectReference Include="..\..\..\src\Stride.CommunityToolkit.Bullet\Stride.CommunityToolkit.Bullet.csproj" />
       <ProjectReference Include="..\..\..\src\Stride.CommunityToolkit.Windows\Stride.CommunityToolkit.Windows.csproj" />
     </ItemGroup>
