+++ examples/code-only/Example01_Basic3DScene/Program.cs
@@ -1,3 +1,4 @@
+using Stride.CommunityToolkit.Bullet;
 using Stride.CommunityToolkit.Engine;
 using Stride.CommunityToolkit.Rendering.ProceduralModels;
 using Stride.CommunityToolkit.Skyboxes;
