+++ examples/snippets/Easing_Example01/Program.cs
@@ -1,3 +1,4 @@
+using Stride.CommunityToolkit.Bullet;
 using Stride.CommunityToolkit.Engine;
 using Stride.CommunityToolkit.Mathematics;
 using Stride.CommunityToolkit.Rendering.ProceduralModels;
