+++ examples/snippets/Physics_Example001/Program.cs
@@ -1,3 +1,4 @@
+using Stride.CommunityToolkit.Bullet;
 using Stride.CommunityToolkit.Engine;
 using Stride.CommunityToolkit.Games;
 using Stride.CommunityToolkit.Physics;
