+++ src/Stride.CommunityToolkit.Bepu/Stride.CommunityToolkit.Bepu.csproj
@@ -1,9 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
-        <Title>Stride Community Toolkit - Bepu</Title>
+        <Title>Stride Community Toolkit - Bepu Physics</Title>
         <Description>
-            A specialized extension of the Stride Community Toolkit, providing easy-to-use Bepu physics utilities for Stride projects. Enhance your Stride scenes with realistic physics simulations without needing additional assets or complex setup.
+            A specialized extension of the Stride Community Toolkit, providing easy-to-use Bepu Physics utilities for Stride projects. Enhance your Stride scenes with realistic physics simulations without needing additional assets or complex setup.
         </Description>
     </PropertyGroup>
